+++ src/data/sources/mongodb-data-source.ts
@@ -1,22 +1,26 @@
 import { DataSource } from '@src/interfaces/database/data-source'
 import { DefaultResponse } from '@src/interfaces/database/default-response'
 import { MongoDbWrapper } from '@src/interfaces/database/mongodb-wrapper'
+import { QueryFields } from '@src/interfaces/database/query-fields'
 import { MovieBase, MovieResponse } from '@src/models/Movie'
+import { UserBase, UserResponse } from '@src/models/User'
+import { transformObjectId } from '@src/utils/helper'
 
 export class MongoDbDataSource implements DataSource {
     constructor(private db: MongoDbWrapper) {}
 
-    async find<T extends MovieResponse>(query: object, search?: string): Promise<DefaultResponse<Array<T>>> {
+    async find(query: QueryFields, search?: string): Promise<DefaultResponse<any>> {
+        const { id, uid, ...otherQueries } = query
+        if (id && !uid) {
+            query = { uid: id, ...otherQueries }
+        }
         const queryString = search ? { ...query, $text: { $search: search } } : query
         const results = await this.db.find(queryString)
 
         if (results.length)
             return {
                 acknowledged: true,
-                data: results.map(item => {
-                    const { _id, ...other } = item
-                    return { id: _id, ...other }
-                }),
+                data: results,
                 error: null,
             }
         return {
@@ -26,7 +30,9 @@ export class MongoDbDataSource implements DataSource {
         }
     }
 
-    async insertOne<T extends MovieBase, U extends MovieResponse>(doc: T): Promise<DefaultResponse<U>> {
+    async insertOne<T extends MovieBase | UserBase, U extends MovieResponse | UserResponse>(
+        doc: T
+    ): Promise<DefaultResponse<U>> {
         const { acknowledged } = await this.db.insertOne(doc)
         return {
             acknowledged,
@@ -35,7 +41,7 @@ export class MongoDbDataSource implements DataSource {
         }
     }
 
-    async findOneByIdAndUpdate<T extends MovieBase, U extends MovieResponse>(
+    async findOneByIdAndUpdate<T extends MovieBase | UserBase, U extends MovieResponse | UserResponse>(
         id: string,
         update: T
     ): Promise<DefaultResponse<U>> {
@@ -59,8 +65,11 @@ export class MongoDbDataSource implements DataSource {
         return response
     }
 
-    async findOneByIdAndDelete<T extends MovieResponse>(id: string): Promise<DefaultResponse<T>> {
-        const { acknowledged, deletedCount } = await this.db.deleteOne(id)
+    async findOneByIdAndDelete<T extends MovieResponse | UserResponse>(
+        id: string,
+        uid: string
+    ): Promise<DefaultResponse<T>> {
+        const { acknowledged, deletedCount } = await this.db.deleteOne(id, uid)
         return {
             acknowledged: acknowledged && !!deletedCount,
             data: null,
